Extracted finder and attribute methods for Model::Persistence into separate files/modules
