+++ lib/slingshot.rb
@@ -20,6 +20,8 @@ require 'slingshot/dsl'
 require 'slingshot/model/naming'
 require 'slingshot/model/callbacks'
 require 'slingshot/model/search'
+require 'slingshot/model/persistence/finders'
+require 'slingshot/model/persistence/attributes'
 require 'slingshot/model/persistence'
 
 module Slingshot
